+++ build_free5GC.sh
@@ -10,14 +10,19 @@ cd $workdir
 echo "Start to configure free5GC on $node_name ..."
 
 echo "========= Check whether GO is installed ========="
-if command -v go >/dev/null 2>&1; then 
-    echo 'exists go, remove the existing version and install Go 1.15.8:'
-    # this assumes your current version of Go is in the default location
-    sudo rm -rf /usr/local/go
-    wget https://dl.google.com/go/go1.15.8.linux-amd64.tar.gz
-    sudo tar -C /usr/local -zxvf go1.15.8.linux-amd64.tar.gz
-    rm go1.15.8.linux-amd64.tar.gz
-else 
+if command -v go >/dev/null 2>&1; then
+    version=$(go version | awk '{print $3}')
+    if [ $version == 'go1.15.8' ]; then
+        echo "Current go version is $version"
+    else
+        echo 'exists go, remove the existing version and install Go 1.15.8:'
+        # this assumes your current version of Go is in the default location
+        sudo rm -rf /usr/local/go
+        wget https://dl.google.com/go/go1.15.8.linux-amd64.tar.gz
+        sudo tar -C /usr/local -zxvf go1.15.8.linux-amd64.tar.gz
+        rm go1.15.8.linux-amd64.tar.gz
+    fi
+else
     echo 'no exists go'
     wget https://dl.google.com/go/go1.15.8.linux-amd64.tar.gz
     sudo tar -C /usr/local -zxvf go1.15.8.linux-amd64.tar.gz
